整理程式碼
